virtual_ihc_diffusion/requirements.txt
# Virtual IHC Staining - Requirements
# Python 3.10+

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# MONAI
monai[all]>=1.4.0

# Training & Logging
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=6.0

# Image Processing
scikit-image>=0.21.0
opencv-python>=4.8.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
